Move functions which can be moved into bilininteg_*_kernels.cpp files in order to speed up compilation
